refactor(extraction): keep grounding wrapper under ratchet
