+++ packages/remnic-core/src/extraction-source-grounding.ts
@@ -1,4 +1,5 @@
 import type { ExtractionResult } from "./types.js";
+import { delinearize } from "./delinearize.js";
 import {
   applyExtractionSourceGrounding as applyGroundingRules,
   filterExtractionResultBySource as filterGroundingRules,
@@ -18,11 +19,21 @@ export function applyExtractionSourceGrounding(
   messageTimestamp: Date | undefined,
   options: ExtractionSourceGroundingOptions,
 ): ExtractionResult {
+  const sourceForGrounding = (source: string | undefined): string | undefined =>
+    source === undefined || !options.anchorTemporalExpressions
+      ? source
+      : delinearize(source, result.entities, messageTimestamp ?? new Date());
+  const groundedRoleSources = roleAssertionSources === undefined
+    ? undefined
+    : {
+      profile: sourceForGrounding(roleAssertionSources.profile),
+      identity: sourceForGrounding(roleAssertionSources.identity),
+    };
   return applyGroundingRules(
     result,
-    sourceText,
-    assertionSourceText,
-    roleAssertionSources,
+    sourceForGrounding(sourceText) ?? sourceText,
+    sourceForGrounding(assertionSourceText) ?? assertionSourceText,
+    groundedRoleSources,
     messageTimestamp,
     options,
   );
