+++ packages/remnic-core/src/extraction.ts
@@ -1152,21 +1152,11 @@ export class ExtractionEngine {
     messageTimestamp?: Date,
   ): ExtractionResult {
     const capabilities = resolvePipelineProcessingCapabilities(this.config);
-    const sourceForGrounding = (source: string | undefined): string | undefined =>
-      source === undefined || !capabilities.delinearize
-        ? source
-        : delinearize(source, result.entities, messageTimestamp ?? new Date());
-    const groundedRoleSources = roleAssertionSources === undefined
-      ? undefined
-      : {
-        profile: sourceForGrounding(roleAssertionSources.profile),
-        identity: sourceForGrounding(roleAssertionSources.identity),
-      };
     return applyExtractionSourceGrounding(
       result,
-      sourceForGrounding(sourceText) ?? sourceText,
-      sourceForGrounding(assertionSourceText) ?? assertionSourceText,
-      groundedRoleSources,
+      sourceText,
+      assertionSourceText,
+      roleAssertionSources,
       messageTimestamp,
       {
         sourceGrounding: capabilities.sourceGrounding,
